Create pytest.ini
we will  need pytest.ini file for configuration of Pytest:

The first thing we do here is set a bunch of command line arguments — we enable colors in terminal output, then we enable coverage reporting for blueprint directory, after that we enable both generations of XML and stdout ( term) coverage reports. Final 2 arguments (-ra) tell Pytest to output short summary for non-passing tests.
On the next line, we have filterwarnings option which allows us to disable some annoying warnings in the output, for example, deprecation warnings coming out of some library which we have no control over.
Rest of the config sets up logging. First one just turns it on and other 3 configure level, format and datetime format. Easier than explaining the format config is just seeing the output itself, which is shown in the next section.
With all the configurations in pytest.ini, all we will need to do to run our test suite is run pytest, not even the package argument needed!