Create requirements.txt
Last  configuration file we have is requirement.txt, which contains a list of our dependencies. All you can find in this file is a list of Python packages, one per line with the optional version of the package. As noted, the package version is optional, but we strongly suggest you lock versions in requirements.txt to avoid situations, where you might download newer, incompatible package during build and deployment, and end-up breaking your application.